requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymysql==1.0.2
python-dotenv==0.19.1